Added error handling for writing to heartbeat file. (left (io($opt{disable_alerts})->touch alone, because at worst, it keeps warning)
Added "alert" and "notify" commandline options.  These are overridden
by config file lines, but are handy for testing.

Added -alert $(USER) to Makefile tests to pass current user to the
tests to check email warnings.

Modified tests to use "perl task_schedule.pl" to allow the $SKA env
to actually be used for testing.  task_schedule.pl is still set with
and explicit path shebang and probably should be.