Add task_sched_no_master_heart_attack file to override master_heart_attack
Certain tasks should attempt to keep running even if there is a
master_heart_attack signal set.  This can now be done by touching a file
named "task_sched_no_master_heart_attack" in the task data directory.

Testing of this patch consisted of starting the arc task in a Ska dev
environment with master_heart_attack set and verifying that the task
did not start.  Then task_sched_no_master_heart_attack was touched to
confirm that the task ran normally.  Finally task_sched_no_master_heart_attack
was removed and it was confirmed that the task was stopped as expected.